requirements.txt
astroid==2.1.0
autopep8==1.4.3
certifi==2017.11.5
chardet==3.0.4
entrypoints==0.3
flake8==3.7.5
gitdb2==2.0.3
GitPython==2.1.7
idna==2.6
isort==4.3.4
lazy-object-proxy==1.3.1
Markdown==2.6.10
mccabe==0.6.1
olefile==0.44
Pillow==5.4.1
pycodestyle==2.5.0
pyflakes==2.1.0
PyJWT==1.7.1
pylint==2.2.2
six==1.12.0
smmap2==2.0.3
urllib3==1.24.1
wrapt==1.11.1

Django==1.11.17
django-filter==1.1.0
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
pytz==2018.9
requests==2.21.0